+++ src/utils.js
@@ -39,6 +39,41 @@ function isPlanB() {
   return browser() === 'chrome' || browser() === 'safari';
 }
 
+function enabledSimulcast(role, video) {
+  /**
+    VP9 x
+
+    simulcast_pub Chrome o
+    simulcast_pub Firefox x
+    simulcast_pub Safari 12.1 o
+    simulcast_pub Safari 12.0 x
+    simulcast_sub Chrome o
+    simulcast_sub Firefox o
+    simulcast_sub Safari 12.1 o
+    simulcast_sub Safari 12.0 o ※H.264 のみ
+  **/
+  if (video.codec_type === 'VP9') {
+    return false;
+  }
+  if (role === 'upstream' && browser() === 'firefox') {
+    return false;
+  }
+  if (browser() === 'safari') {
+    const appVersion = window.navigator.appVersion.toLowerCase();
+    const version = /version\/([\d.]+)/.exec(appVersion).pop();
+    // version 12.0 以降であれば有効
+    if (12.0 < parseFloat(version)) {
+      return true;
+    }
+    if (12.0 == parseFloat(version) && role === 'downstream' && video.codec_type === 'H264') {
+      // role が downstream で 'H264' の場合のみ有効
+      return true;
+    }
+    return false;
+  }
+  return true;
+}
+
 export function isUnifiedChrome() {
   if (browser() !== 'chrome') {
     return false;
@@ -51,13 +86,11 @@ export function isUnifiedChrome() {
   return 71 <= parseInt(splitedUserAgent[1]);
 }
 
-export function isUnifiedSafari() {
+export function isUnifiedSafari(sdp) {
   if (browser() !== 'safari') {
     return false;
   }
-  const appVersion = window.navigator.appVersion.toLowerCase();
-  const version = /version\/([\d.]+)/.exec(appVersion).pop();
-  return 12.0 < parseFloat(version);
+  return sdp.includes('a=group:BUNDLE 0 1');
 }
 
 export function isEdge() {
@@ -111,7 +144,7 @@ export function createSignalingMessage(offerSDP, role, channelId, metadata, opti
   if ('multistream' in options && options.multistream === true) {
     // multistream
     message.multistream = true;
-    if (!isUnifiedChrome() && !isUnifiedSafari() && isPlanB()) {
+    if (!isUnifiedChrome() && !isUnifiedSafari(offerSDP) && isPlanB()) {
       message.plan_b = true;
     }
     // spotlight
@@ -119,9 +152,6 @@ export function createSignalingMessage(offerSDP, role, channelId, metadata, opti
       message.spotlight = options.spotlight;
     }
   } else if ('simulcast' in options || 'simulcastQuality' in options) {
-    if (!(isUnifiedSafari() || isChrome())) {
-      throw new Error('Simulcast can not be used with this browse be used with this browserr');
-    }
     // simulcast
     if ('simulcast' in options && options.simulcast === true) {
       message.simulcast = true;
@@ -177,5 +207,8 @@ export function createSignalingMessage(offerSDP, role, channelId, metadata, opti
     }
   }
 
+  if (message.simulcast && !enabledSimulcast(message.role, message.video)) {
+    throw new Error('Simulcast can not be used with this browser');
+  }
   return message;
 }