+++ test/utils.spec.js
@@ -254,7 +254,6 @@ describe('Utils', () => {
           metadata: metadata,
           multistream: true,
           sdp: sdp,
-          plan_b: true,
           video: true,
           audio: true,
           userAgent: userAgent
@@ -280,7 +279,6 @@ describe('Utils', () => {
           audio: true,
           userAgent: userAgent,
           multistream: true,
-          plan_b: true,
           spotlight: 2
         };
         assert.deepEqual(actual, expected);
@@ -363,6 +361,97 @@ describe('Utils', () => {
         };
         assert.deepEqual(actual, expected);
       });
+
+      it('simulcast with VP9', () => {
+        const options = {
+          simulcast: true,
+          videoCodecType: 'VP9',
+        };
+        assert.throws(
+          () => {
+            createSignalingMessage(sdp, 'downstream', channelId, metadata, options);
+          },
+          (error) => {
+            assert(error.message === 'Simulcast can not be used with this browser');
+            return true;
+          }
+        );
+      });
+
+      it('simulcast with Firefox upstream', () => {
+        Object.defineProperty(navigator, 'userAgent', {
+          get: () => { return '"Mozilla/5.0 (Macintosh; Intel Mac OS X 10.14; rv:66.0) Gecko/20100101 Firefox/66.0"'; },
+          enumerable: true,
+          configurable: true,
+        });
+        const options = {
+          simulcast: true
+        };
+        assert.throws(
+          () => {
+            createSignalingMessage(sdp, 'upstream', channelId, metadata, options);
+          },
+          (error) => {
+            assert(error.message === 'Simulcast can not be used with this browser');
+            return true;
+          }
+        );
+      });
+
+      it('simulcast with Safari 12.0 upstream', () => {
+        const localUserAgent = '"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_4) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/12.0 Safari/605.1.15"';  // eslint-disable-line
+        Object.defineProperty(navigator, 'userAgent', {
+          get: () => { return localUserAgent; },
+          enumerable: true,
+          configurable: true,
+        });
+        const appVersion = '"5.0 (Macintosh; Intel Mac OS X 10_14_4) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/12.0 Safari/605.1.15"';  // eslint-disable-line
+        Object.defineProperty(navigator, 'appVersion', {
+          get: () => { return appVersion; },
+          enumerable: true,
+          configurable: true,
+        });
+        const options = {
+          simulcast: true
+        };
+        assert.throws(
+          () => {
+            createSignalingMessage(sdp, 'upstream', channelId, metadata, options);
+          },
+          (error) => {
+            assert(error.message === 'Simulcast can not be used with this browser');
+            return true;
+          }
+        );
+      });
+
+      it('simulcast with Safari 12.0 VP8 downstream', () => {
+        const localUserAgent = '"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_4) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/12.0 Safari/605.1.15"';  // eslint-disable-line
+        Object.defineProperty(navigator, 'userAgent', {
+          get: () => { return localUserAgent; },
+          enumerable: true,
+          configurable: true,
+        });
+        const appVersion = '"5.0 (Macintosh; Intel Mac OS X 10_14_4) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/12.0 Safari/605.1.15"';  // eslint-disable-line
+        Object.defineProperty(navigator, 'appVersion', {
+          get: () => { return appVersion; },
+          enumerable: true,
+          configurable: true,
+        });
+        const options = {
+          simulcast: true
+        };
+        assert.throws(
+          () => {
+            createSignalingMessage(sdp, 'downstream', channelId, metadata, options);
+          },
+          (error) => {
+            assert(error.message === 'Simulcast can not be used with this browser');
+            return true;
+          }
+        );
+      });
+
     });
   });
 });